Use requiredMods property instead of dependencies
The dependencies were not being parsed properly, resulting in no dependencies. The requiredMods property unfortunately does not support side arguments (or any arguments), so CTM is now required by both client and server. Sided dependencies can be confusing to users though, anyway, so it's not the end of the world.