Remove use of MarkupBuilder
Groovy's XML package is no longer native in newer Gradle versions so
 this is simpler for future migrations.

Managing all tags directly also allows for a cleaner implementation of unordered list scopes.

The output is now minified as new lines and indentation are not necessary for the changelog markup.

This commit also updates the variables declared at the top of the buildscript to be global.